requirements.txt
requests==2.31.0
python-dotenv==1.0.0
numpy==1.25.2
pydantic==1.10.12
sqlalchemy==2.0.20
fastapi==0.99.1
tenacity==8.2.3
openai==0.27.9
tiktoken==0.4.0
langchain==0.0.273
uvicorn==0.23.2
beautifulsoup4==4.12.2
pinecone-client==2.2.2